requirements.txt
djangorestframework-jwt
aliyun-python-sdk-cdn==3.3.0
aliyun-python-sdk-core==2.13.15
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-ecs==4.19.1
asgiref==3.2.5
Django==3.0.4
django-filter==2.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
djangorestframework==3.11.0
jmespath==0.9.5
mysqlclient==1.4.6
Pillow==7.1.1
pycryptodome==3.9.7
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
